Removed definition of URL for "qt" namespace from individual parts of the model, and instead define it once at the top.
This is the way that Bob's tools generate a v3 sys defn, and it brings the auto generated one in sync with the one usable in the build.

Also added a v3 system definition for S^4, although it's not quite in sync with the v2 sys defn there (because that's not in sync with the generated sys defns at the moment).